+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchScreen.kt
@@ -1,24 +1,23 @@
 package org.simple.clinic.instantsearch
 
-import android.annotation.SuppressLint
 import android.content.Context
-import android.os.Parcelable
-import android.util.AttributeSet
+import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
-import android.view.WindowManager.LayoutParams.SOFT_INPUT_ADJUST_NOTHING
-import android.view.WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE
+import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
 import androidx.appcompat.app.AppCompatActivity
-import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.RecyclerView
 import com.jakewharton.rxbinding3.recyclerview.scrollStateChanges
 import com.jakewharton.rxbinding3.view.clicks
-import com.jakewharton.rxbinding3.view.detaches
 import com.jakewharton.rxbinding3.widget.editorActions
 import com.jakewharton.rxbinding3.widget.textChanges
 import io.reactivex.Observable
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.disposables.Disposable
+import io.reactivex.rxkotlin.cast
 import io.reactivex.rxkotlin.ofType
-import kotlinx.android.synthetic.main.screen_instant_search.view.*
+import kotlinx.android.synthetic.main.screen_instant_search.*
 import org.simple.clinic.R
 import org.simple.clinic.bp.assignbppassport.BpPassportSheet
 import org.simple.clinic.databinding.ListPatientSearchBinding
@@ -28,10 +27,9 @@ import org.simple.clinic.di.injector
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.facility.alertchange.AlertFacilityChangeSheet
 import org.simple.clinic.facility.alertchange.Continuation
-import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.navigation.v2.Router
 import org.simple.clinic.navigation.v2.compat.wrap
-import org.simple.clinic.navigation.v2.keyprovider.ScreenKeyProvider
+import org.simple.clinic.navigation.v2.fragments.BaseScreen
 import org.simple.clinic.newentry.PatientEntryScreenKey
 import org.simple.clinic.patient.PatientSearchResult
 import org.simple.clinic.patient.businessid.Identifier
@@ -42,7 +40,6 @@ import org.simple.clinic.summary.OpenIntention
 import org.simple.clinic.summary.PatientSummaryScreenKey
 import org.simple.clinic.util.UtcClock
 import org.simple.clinic.util.extractSuccessful
-import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.widgets.ItemAdapter
 import org.simple.clinic.widgets.UiEvent
 import org.simple.clinic.widgets.hideKeyboard
@@ -52,7 +49,16 @@ import java.util.UUID
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
-class InstantSearchScreen(context: Context, attrs: AttributeSet) : ConstraintLayout(context, attrs), InstantSearchUi, InstantSearchUiActions {
+class InstantSearchScreen :
+    BaseScreen<
+        InstantSearchScreenKey,
+        ScreenInstantSearchBinding,
+        InstantSearchModel,
+        InstantSearchEvent,
+        InstantSearchEffect,
+        InstantSearchUiRenderer>(),
+    InstantSearchUi,
+    InstantSearchUiActions {
 
   companion object {
     private const val BP_PASSPORT_SHEET = 1333
@@ -74,32 +80,25 @@ class InstantSearchScreen(context: Context, attrs: AttributeSet) : ConstraintLay
   @Inject
   lateinit var utcClock: UtcClock
 
-  @Inject
-  lateinit var screenKeyProvider: ScreenKeyProvider
-
-  private var binding: ScreenInstantSearchBinding? = null
+  private val subscriptions = CompositeDisposable()
 
   private val instantSearchToolbar
-    get() = binding!!.instantSearchToolbar
+    get() = binding.instantSearchToolbar
 
   private val searchQueryEditText
-    get() = binding!!.searchQueryEditText
+    get() = binding.searchQueryEditText
 
   private val searchQueryTextInputLayout
-    get() = binding!!.searchQueryTextInputLayout
+    get() = binding.searchQueryTextInputLayout
 
   private val searchResultsView
-    get() = binding!!.searchResultsView
+    get() = binding.searchResultsView
 
   private val newPatientButton
-    get() = binding!!.newPatientButton
+    get() = binding.newPatientButton
 
   private val instantSearchProgressIndicator
-    get() = binding!!.instantSearchProgressIndicator
-
-  private val screenKey: InstantSearchScreenKey by unsafeLazy {
-    screenKeyProvider.keyFor(this)
-  }
+    get() = binding.instantSearchProgressIndicator
 
   private val allPatientsAdapter = ItemAdapter(
       diffCallback = InstantSearchResultsItemType.DiffCallback(),
@@ -125,64 +124,38 @@ class InstantSearchScreen(context: Context, attrs: AttributeSet) : ConstraintLay
       )
   )
 
-  private val events by unsafeLazy {
-    Observable.mergeArray(
-        allPatientsItemClicks(),
-        searchItemClicks(),
-        searchQueryChanges(),
-        registerNewPatientClicks(),
-        blankBpPassportResults()
-    )
-  }
+  override fun defaultModel() = InstantSearchModel.create(screenKey.additionalIdentifier)
 
-  private val delegate by unsafeLazy {
-    val uiRenderer = InstantSearchUiRenderer(this)
+  override fun uiRenderer() = InstantSearchUiRenderer(this)
 
-    MobiusDelegate.forView(
-        events = events.ofType(),
-        defaultModel = InstantSearchModel.create(screenKey.additionalIdentifier),
-        init = InstantSearchInit(),
-        update = InstantSearchUpdate(),
-        effectHandler = effectHandlerFactory.create(this).build(),
-        modelUpdateListener = uiRenderer::render
-    )
-  }
+  override fun bindView(layoutInflater: LayoutInflater, container: ViewGroup?) =
+      ScreenInstantSearchBinding.inflate(layoutInflater, container, false)
 
-  private val detaches = detaches()
+  override fun events() = Observable
+      .mergeArray(
+          allPatientsItemClicks(),
+          searchItemClicks(),
+          searchQueryChanges(),
+          registerNewPatientClicks(),
+          blankBpPassportResults()
+      ).cast<InstantSearchEvent>()
 
-  override fun onAttachedToWindow() {
-    super.onAttachedToWindow()
-    delegate.start()
-  }
+  override fun createUpdate() = InstantSearchUpdate()
 
-  override fun onDetachedFromWindow() {
-    delegate.stop()
-    binding = null
+  override fun createInit() = InstantSearchInit()
 
-    activity.window.setSoftInputMode(SOFT_INPUT_ADJUST_RESIZE)
+  override fun createEffectHandler() = effectHandlerFactory.create(this).build()
 
-    super.onDetachedFromWindow()
-  }
-
-  override fun onSaveInstanceState(): Parcelable {
-    return delegate.onSaveInstanceState(super.onSaveInstanceState())
-  }
-
-  override fun onRestoreInstanceState(state: Parcelable?) {
-    super.onRestoreInstanceState(delegate.onRestoreInstanceState(state))
-  }
-
-  override fun onFinishInflate() {
-    super.onFinishInflate()
-    if (isInEditMode) return
+  override fun onAttach(context: Context) {
+    super.onAttach(context)
     context.injector<Injector>().inject(this)
+  }
 
-    activity.window.setSoftInputMode(SOFT_INPUT_ADJUST_NOTHING)
-
-    binding = ScreenInstantSearchBinding.bind(this)
+  override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+    super.onViewCreated(view, savedInstanceState)
 
     // When a scanned BP Passport does not result in a match, we bring up a bottom sheet which asks
-    // whether this is a new registration or an existing patient. If we show the keyboard in these 
+    // whether this is a new registration or an existing patient. If we show the keyboard in these
     // cases, the UI is janky since the keyboard pops up and immediately another bottom sheet pops up.
     // This improves the experience by showing the keyboard only if we have arrived here by searching
     // for a patient by the name
@@ -195,9 +168,16 @@ class InstantSearchScreen(context: Context, attrs: AttributeSet) : ConstraintLay
 
     searchResultsView.adapter = allPatientsAdapter
 
-    setupAlertResults()
-    hideKeyboardOnSearchResultsScroll()
-    hideKeyboardOnImeAction()
+    subscriptions.addAll(
+        setupAlertResults(),
+        hideKeyboardOnSearchResultsScroll(),
+        hideKeyboardOnImeAction()
+    )
+  }
+
+  override fun onDestroyView() {
+    super.onDestroyView()
+    subscriptions.clear()
   }
 
   override fun showAllPatients(patients: List<PatientSearchResult>, facility: Facility) {
@@ -233,42 +213,32 @@ class InstantSearchScreen(context: Context, attrs: AttributeSet) : ConstraintLay
   }
 
   override fun openBpPassportSheet(identifier: Identifier) {
-    val intent = BpPassportSheet.intent(context, identifier)
+    val intent = BpPassportSheet.intent(requireContext(), identifier)
     activity.startActivityForResult(intent, BP_PASSPORT_SHEET)
   }
 
   override fun showNoPatientsInFacility(facility: Facility) {
     searchResultsView.visibility = View.GONE
     noPatientsInFacilityContainer.visibility = View.VISIBLE
-    noPatientsInFacilityTextView.text = context.getString(R.string.patientsearch_error_no_patients_in_facility_heading, facility.name)
-
-    // In order to get the create button above the keyboard, we
-    // update the soft input mode.
-    activity.window.setSoftInputMode(SOFT_INPUT_ADJUST_RESIZE)
+    noPatientsInFacilityTextView.text = getString(R.string.patientsearch_error_no_patients_in_facility_heading, facility.name)
   }
 
   override fun showNoSearchResults() {
     searchResultsView.visibility = View.GONE
     noSearchResultsContainer.visibility = View.VISIBLE
-
-    // In order to get the create button above the keyboard, we
-    // update the soft input mode.
-    activity.window.setSoftInputMode(SOFT_INPUT_ADJUST_RESIZE)
   }
 
   override fun hideNoPatientsInFacility() {
     noPatientsInFacilityContainer.visibility = View.GONE
-    activity.window.setSoftInputMode(SOFT_INPUT_ADJUST_NOTHING)
   }
 
   override fun hideNoSearchResults() {
     noSearchResultsContainer.visibility = View.GONE
-    activity.window.setSoftInputMode(SOFT_INPUT_ADJUST_NOTHING)
   }
 
   override fun openPatientEntryScreen(facility: Facility) {
     activity.startActivityForResult(
-        AlertFacilityChangeSheet.intent(context, facility.name, Continuation.ContinueToScreen(PatientEntryScreenKey())),
+        AlertFacilityChangeSheet.intent(requireContext(), facility.name, Continuation.ContinueToScreen(PatientEntryScreenKey())),
         ALERT_FACILITY_CHANGE
     )
   }
@@ -281,15 +251,13 @@ class InstantSearchScreen(context: Context, attrs: AttributeSet) : ConstraintLay
     instantSearchProgressIndicator.visibility = View.GONE
   }
 
-  @SuppressLint("CheckResult")
-  private fun setupAlertResults() {
-    screenResults
+  private fun setupAlertResults(): Disposable {
+    return screenResults
         .streamResults()
         .ofType<ActivityResult>()
         .extractSuccessful(ALERT_FACILITY_CHANGE) { intent ->
           AlertFacilityChangeSheet.readContinuationExtra<Continuation.ContinueToScreen>(intent).screenKey
         }
-        .takeUntil(detaches)
         .map(FullScreenKey::wrap)
         .subscribe(router::push)
   }
@@ -329,21 +297,17 @@ class InstantSearchScreen(context: Context, attrs: AttributeSet) : ConstraintLay
         .map(::BlankBpPassportResultReceived)
   }
 
-  @SuppressLint("CheckResult")
-  private fun hideKeyboardOnSearchResultsScroll() {
-    searchResultsView
+  private fun hideKeyboardOnSearchResultsScroll(): Disposable {
+    return searchResultsView
         .scrollStateChanges()
         .filter { it == RecyclerView.SCROLL_STATE_DRAGGING }
-        .takeUntil(detaches)
-        .subscribe { hideKeyboard() }
+        .subscribe { binding.root.hideKeyboard() }
   }
 
-  @SuppressLint("CheckResult")
-  private fun hideKeyboardOnImeAction() {
-    searchQueryEditText
+  private fun hideKeyboardOnImeAction(): Disposable {
+    return searchQueryEditText
         .editorActions { actionId -> actionId == EditorInfo.IME_ACTION_SEARCH }
-        .takeUntil(detaches)
-        .subscribe { hideKeyboard() }
+        .subscribe { binding.root.hideKeyboard() }
   }
 
   interface Injector {