+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchScreenKey.kt
@@ -3,19 +3,16 @@ package org.simple.clinic.instantsearch
 import android.os.Parcelable
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
+import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.patient.businessid.Identifier
-import org.simple.clinic.router.screen.FullScreenKey
 
 @Parcelize
 data class InstantSearchScreenKey(
     val additionalIdentifier: Identifier?
-) : FullScreenKey, Parcelable {
+) : ScreenKey(), Parcelable {
 
   @IgnoredOnParcel
   override val analyticsName: String = "Instant Search Screen"
 
-  override fun layoutRes(): Int {
-    return R.layout.screen_instant_search
-  }
+  override fun instantiateFragment() = InstantSearchScreen()
 }